vertex-ai/requirements-training.txt
prophet==1.1.5
pandas==2.0.3
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.0
